+++ tuk-api/src/main/kotlin/nexters/tuk/application/push/PushService.kt
@@ -50,7 +50,7 @@ class PushService(
                         memberNameMap[token.memberId]
                             ?: return@forEach
                     ),
-                    body = pushMessage.body
+                    body = pushMessage.body,
                 )
             )
         }
