+++ tuk-api/src/main/kotlin/nexters/tuk/ui/gathering/GatheringController.kt
@@ -1,9 +1,10 @@
 package nexters.tuk.ui.gathering
 
 import nexters.tuk.application.gathering.GatheringCommandService
-import nexters.tuk.application.gathering.GatheringGenerateService
+import nexters.tuk.application.gathering.GatheringService
 import nexters.tuk.application.gathering.GatheringMemberService
 import nexters.tuk.application.gathering.GatheringQueryService
+import nexters.tuk.application.gathering.dto.request.GatheringCommand
 import nexters.tuk.application.gathering.dto.request.GatheringQuery
 import nexters.tuk.application.gathering.dto.response.GatheringMemberResponse
 import nexters.tuk.application.gathering.dto.response.GatheringResponse
@@ -14,8 +15,7 @@ import org.springframework.web.bind.annotation.*
 @RestController
 @RequestMapping("/api/v1/gatherings")
 class GatheringController(
-    private val gatheringCommandService: GatheringCommandService,
-    private val gatheringGenerateService: GatheringGenerateService,
+    private val gatheringService: GatheringService,
     private val gatheringQueryService: GatheringQueryService,
     private val gatheringMemberService: GatheringMemberService,
 ) : GatheringSpec {
@@ -26,7 +26,7 @@ class GatheringController(
         @RequestBody request: GatheringDto.Request.Generate,
     ): ApiResponse<GatheringResponse.Generate> {
 
-        val response = gatheringGenerateService.generateGathering(request.toCommand(memberId))
+        val response = gatheringService.generateGathering(request.toCommand(memberId))
 
         return ApiResponse.ok(response)
     }
@@ -37,7 +37,12 @@ class GatheringController(
         @PathVariable gatheringId: Long,
         @RequestBody request: GatheringDto.Request.Update,
     ): ApiResponse<GatheringResponse.Simple> {
-        val response = gatheringCommandService.updateGathering(request.toCommand(gatheringId))
+        val response = gatheringService.updateGathering(
+            GatheringCommand.Update(
+                gatheringId = gatheringId,
+                gatheringIntervalDays = request.gatheringIntervalDays
+            )
+        )
 
         return ApiResponse.ok(response)
     }
@@ -78,7 +83,7 @@ class GatheringController(
 
     @GetMapping("/{gatheringId}/name")
     override fun getGatheringName(
-        @PathVariable("gatheringId") gatheringId: Long
+        @PathVariable("gatheringId") gatheringId: Long,
     ): ApiResponse<GatheringResponse.GatheringName> {
 
         val response = gatheringQueryService.getGatheringName(gatheringId)
